demo/requirements.txt
streamlit
langchain==0.1.3
openai==1.9.0
loguru==0.7.2
python-dotenv